obsolete(XRITDialog, XRITControllerManager): Moved to XRIT module and marked as obsolete.
